finessing the break choice viz, adding some comments
