+++ public/choice-bar.html
@@ -5,7 +5,7 @@
         <script src="js/vendor/jquery-1.12.4.min.js"></script>
         <script src="js/vendor/d3.v4.min.js"></script>
         <script src="js/vendor/firebase-3.6.1.js"></script>
-        <script src="js/viz.bundle.js"></script>
+        <script src="js/viz.bar.bundle.js"></script>
         <meta name="viewport" content="width=device-width, initial-scale=1.0, minimum-scale=1.0">
     </head>
     <body>
@@ -14,7 +14,7 @@
             <div class="clock" id="clock"></div>
         </div>
 
-        <svg class="viz"></svg>
+        <div class="viz"></div>
 
         <script>
         /*global firebase, jQuery*/
@@ -31,9 +31,7 @@
 
         firebase.initializeApp(config);
         preferences = firebase.database().ref('presenters/');
-        preferences.on('value', function(snapshot) {
-            drawClusters(voteCounts(snapshot.val()));
-        });
+        preferences.on('value', drawBarGraph);
         </script>
         <script src="js/vendor/time.js"></script>
     </body>
