+++ public/css/viz.css
@@ -1,7 +1,8 @@
 body {
     background-color: #3b4150;
-    font-family: arial, sans-serif;
+    font-family: CNN, 'Helvetica Neue', Helvetica, arial, sans-serif;
     color: #cdcdcd;
+    margin: 0;
 }
 
 .viz {
@@ -10,20 +11,32 @@ body {
     margin: 0 auto;
 }
 
+.date {
+    padding: .5rem 1rem;
+    font-weight: 100;
+    font-size: 1.5rem;
+    max-width: 960px;
+    margin: 0 auto;
+    height: 0;
+    overflow: visible;
+}
+
 /****************************************** BAR GRAPH */
 
 .bar-graph {
     display: flex;
-    justify-content: space-around;
     align-items: flex-end;
     position: relative;
     height: 100%;
     max-height: 60vh;
+    padding: 0 1rem 1rem;
+    border-bottom: 1px solid #707079;
 }
 
 .bar {
     position: relative;
     width: 100%;
+    max-width: 160px;
     background: #707079;
     transition: height .5s;
 }
@@ -35,6 +48,9 @@ body {
     margin-bottom: .5rem;
     left: 50%;
     transform: translate(-50%, 0);
+    font-size: 2.925rem;
+    font-weight: 100;
+    color: #3b4150;
 }
 
 .bar:after {
@@ -43,7 +59,9 @@ body {
     top: 100%;
     left: 50%;
     transform: translate(-50%, 0);
-    margin-top: 1rem;
+    margin-top: 1.5rem;
+    font-size: 1.5rem;
+    font-weight: 100;
 }
 
 .bar + .bar {
@@ -56,4 +74,72 @@ svg.viz {
     width: 100%;
     height: 100%;
     max-width: 100%;
+}
+
+.vote-conan {
+    fill: #369;
+}
+
+.vote-samantha {
+    fill: #c00;
+}
+
+[class*="vote-"].fixed {
+    fill: transparent;
+}
+
+[class*="vote-conan"].fixed {
+    fill: url(#conan-head);
+}
+
+[class*="vote-samantha"].fixed {
+    fill: url(#samantha-head);
+}
+
+/****************************************** BREAK VIZ */
+
+.viz.break {
+    height: 80%;
+}
+
+.viz.break .vote {
+    position: absolute;
+    top: 50%;
+    left: 25%;
+    width: 2rem;
+    height: 2rem;
+    transform: translate(-50%, -50%);
+}
+
+.viz.break .vote-count {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    text-align: center;
+    font-size: 2rem;
+    font-weight: 300;
+    color: #3b4150;
+}
+
+.viz.break .vote[data-label="yes"] .vote-count {
+    margin-top: 2rem;
+}
+
+.viz.break .vote[data-label="no"] .vote-count {
+    margin-top: -2rem;
+}
+
+.viz.break .vote:nth-of-type(2) {
+    left: 75%;
+}
+
+.viz.break .vote[data-label="yes"] {
+    background: url(/img/thumbs-up.png) center center;
+    background-size: contain;
+}
+
+.viz.break .vote[data-label="no"] {
+    background: url(/img/thumbs-down.png) center center;
+    background-size: contain;
 }
